backend/requirements.txt
# ── Core Framework ──
fastapi>=0.103.0
uvicorn[standard]>=0.23.2
pydantic>=2.3.0
pydantic-settings>=2.0.3

# ── Database ──
sqlalchemy[asyncio]>=2.0.20
asyncpg>=0.28.0
psycopg2-binary>=2.9.7
alembic>=1.12.0

# ── Redis ──
redis>=5.0.0

# ── AI / LLM ──
openai>=1.3.0
tiktoken>=0.5.1

# ── Vector / Embeddings ──
numpy>=1.24.0
pgvector>=0.2.0

# ── Auth ──
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ── Utilities ──
httpx>=0.25.0
python-dotenv>=1.0.0
jinja2>=3.1.2
